+++ CHANGELOG.md
@@ -12,6 +12,8 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 
 ### Added
 
+- [#8366](https://github.com/thanos-io/thanos/pull/8366) Store: optionally ignore Parquet migrated blocks
+
 ### Changed
 
 - [#8370](https://github.com/thanos-io/thanos/pull/8370) Query: announced labelset now reflects relabel-config